[Inmethod Grid] Cherry Pick raystorm-inmethod-Grid from 1.5 to master
keep long instead of int in AbstractPageableView.

+++ jdk-1.5-parent/inmethod-grid-parent/inmethod-grid/src/main/java/com/inmethod/grid/common/AbstractPageableView.java
@@ -183,7 +183,7 @@ public abstract class AbstractPageableView<T> extends RefreshingView<T>
 	/**
 	 * Index of the item which is at the beginning on current page
 	 */
-	private int currentPageFirstItem = 0;
+	private long currentPageFirstItem = 0;
 
 	/**
 	 * Cached query result for this request
